pyproject.toml
[tool.poetry]
name = "allcaps"
version = "1.0.0"
description = "Capability-based payload generation framework"
authors = ["2XXE <root@example.com>"]
readme = "README.md"
license = "AGPL-3.0-only"

[tool.poetry.dependencies]
python = "^3.12"
jinja2 = "^3.1.4"
pyyaml = "^6.0.2"
desert = "^2022.9.22"
marshmallow-enum = "^1.5.1"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
allcaps = "allcaps.cli:main"
